Feat: Se agrega comando shell para compilar más rápidamente. Bugfix: Se corrige problema con hebras en Bitonic Sort.

+++ Radix_Sort.cpp
@@ -9,7 +9,7 @@ namespace {
 int getMax(const vector<int>& arr) {
     if (arr.empty()) return 0;
     int maxVal = arr[0];
-    for (int i = 1; i < arr.size(); i++) {
+    for (size_t i = 1; i < arr.size(); i++) {
         if (arr[i] > maxVal) {
             maxVal = arr[i];
         }
